create tracker file
refer to lines 605-644 for the algorithm method. This method then rests inside a another method(single_bot_thread[right above it]) with a  while loop that contintously reads the frames from the camera.  I could use some help implimenting the rest of the algorithm so maybe we could hop on a zoom next week or the week after.

PS this is a small file/class inside a much larger system repository file so allot of parts may not make sense without the rest of the repository but I just wanted to share this part for now.